% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{avaCalib}
\alias{avaCalib}
\title{Calibration for all-vs-all case}
\usage{
avaCalib(
  trnY,
  trnScores,
  tstScores,
  calibMethod,
  deg = NULL,
  K = NULL,
  se = FALSE,
  isProb = NULL
)
}
\arguments{
\item{trnY}{the training labels}

\item{trnScores}{a matrix of decision values or probabilities given by the classifier
for the training sample}

\item{tstScores}{a matrix of decision values or probabilities given by the classifier
for the testing sample}

\item{calibMethod}{a string that indicate which calibration method to use.
options: "knnCalib", "locLinknnCalib", "plattCalib","isoCalib","BBQCalib", "ELiTECalib"}

\item{deg}{an integer to indicate the degree of polynomial for platt scaling}

\item{K}{an integer to indicate the number of nearest neighbor}

\item{isProb}{indicate whether the scores are probabilities}

\item{scaleX}{to determine whether we should standardize the data before using k-NN}

\item{smoothingFtn}{a smoothing function whether it should be applied}
}
\value{
matrix of estimated probabilities for the new cases
}
\description{
a wrapper to implement calibration methods via all-vs-all approach
this is experimental, not fully tested!
}
