% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{bbqCalib}
\alias{bbqCalib}
\title{Calibrate probabilities based on bayesian binning quantile}
\usage{
bbqCalib(trnY, trnScores, tstScores, option = 1)
}
\arguments{
\item{trnScores}{vector/matrix of scores in training set}

\item{tstScores}{scores of new cases option either 1 or 0; averaging=1, selecting=0}

\item{y}{R factor of labels in training set; vector of observed class labels (0/1)}
}
\value{
a vector of probabilities
}
\description{
wrapper calibrate either training scores or probability by
Bayesian Binning
}
