% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{calibWrap}
\alias{calibWrap}
\title{Calibrate probabilities with reliability diagrams}
\usage{
calibWrap(
  trnY,
  tstY,
  trnScores,
  tstScores,
  calibMethod,
  opts = NULL,
  nBins = 25,
  se = FALSE,
  plotsPerRow = 0,
  oneAtATime = TRUE,
  OVA = TRUE,
  isProb = FALSE,
  fitAllX = FALSE,
  smoothingFtn = NULL,
  style = NULL
)
}
\arguments{
\item{trnScores}{vector/matrix of scores output from running the classification method on the training set; will have either c
or c(c-1)/2 columns, where c is the number of classes}

\item{tstScores}{scores for the data to be predicted}

\item{calibMethod}{currently knnCalib or plattCalib}

\item{opts}{R list of classification-specific parameters, e.g.list(k = 50) for knnCalib}

\item{plotsPerRow}{number of plots per row; 0 means no trellis plotting}

\item{oneAtATime}{if TRUE, show the plots one at a time, and give the user the option to print and/or zoom in.}

\item{OVA}{a boolean value to denote whether this is one-vs-all approach if set to False, it will use all-vs-all approach}

\item{isProb}{a boolean value to denote whether scores are probabiltiies}

\item{smoothingFtn}{a smoothing function whether it should be applied. style: 1 or 2 to denote which multicalss reliability diagram style you prefer}
}
\value{
a list that contains a probability matrix

the reliability diagrams for each class
}
\description{
plot the reliability diagrams and compute the calibrated probability matrix for test data.
}
