% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{crossEntropy}
\alias{crossEntropy}
\title{calculates crossEntropy for probability calibrationn algorithms}
\usage{
crossEntropy(calibWrapOut)
}
\arguments{
\item{calibWrapOut}{output object of calibWrap}
}
\value{
cross entropy
}
\description{
calculates crossEntropy for probability calibrationn algorithms
}
