% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{eliteCalib}
\alias{eliteCalib}
\title{Calibration by ELiTe method}
\usage{
eliteCalib(trnY, trnScores, tstScores, build_opt = "AICc", pred_opt = 1)
}
\arguments{
\item{trnScores}{vector of uncalibrated decisions scores for training}

\item{build_opt}{'AIC', or 'AICc' scoring functions (the Elite paper uses AICc).}

\item{pred_opt}{set it to 1  for running model averaging, or to 0 for running model selection}

\item{y}{vector of corresponding true class 1 indicates positive class and 0 indicates negative class.}

\item{tstScore}{vector of uncalibratd decisions scores for testing}
}
\value{
a vector of probabilities
}
\description{
wrapper calibrate probability by ELiTE
}
