% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{fitPlatt}
\alias{fitPlatt}
\title{Fitting a losgitic model}
\usage{
fitPlatt(y, trnScores, deg)
}
\arguments{
\item{y}{the training labels}

\item{trnScores}{a vector of decision values or probabilities given by the classifier for the training sample}

\item{degree}{an integer to indicate the degree of polynomial for the logistic model}
}
\value{
a logistic model
}
\description{
run the logit once and save, rather doing running repeatedly, each
time we have new predictions to make
}
