% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{getCalibMeasures}
\alias{getCalibMeasures}
\title{Wrapper of EliTe error measure for calibration}
\usage{
getCalibMeasures(y, scores)
}
\arguments{
\item{y}{vector of true class of instances {0,1}}

\item{scores}{vector of predictions (classification scores) which is in the closed interval 0 and 1}
}
\value{
a dataframe that shows RMSE, AUROC, ACC, MCE, ECE
}
\description{
provides RMSE, AUROC, ACC, MCE, ECE to measure the calibration performance
}
