% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{getDValsE1071}
\alias{getDValsE1071}
\title{calculcate decision values ("scores") for new cases on previously-fit
e1071 SVM}
\usage{
getDValsE1071(object, newx)
}
\arguments{
\item{object}{the object output by e1071 svm}

\item{newx}{the test set}
}
\value{
decision values based on SVM
}
\description{
for those users of the probability calibration functions on output
from the e1071 package, here are useful utilities
}
