% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{isoCalib}
\alias{isoCalib}
\title{Calibrate probabilities based on isotonic regression}
\usage{
isoCalib(trnY, trnScores, tstScores, isProb)
}
\arguments{
\item{trnY}{the training labels}

\item{trnScores}{the training scores output by the classifier}

\item{tstScores}{the test scores}

\item{isProb}{a boolean value to determine whether scores are probabilities}
}
\value{
a vector of probabilities
}
\description{
wrapper calibrate either training scores or probability by
isotonic regression in binary case
}
