% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{KLDivergence}
\alias{KLDivergence}
\title{calculates Kullback_Leibler divergence for probability calibration}
\usage{
KLDivergence(calibWrapOut)
}
\arguments{
\item{calibWrapOut}{output object of calibWrap}
}
\value{
KL divergence
}
\description{
calculates Kullback_Leibler divergence for probability calibration
}
