% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{knnCalib}
\alias{knnCalib}
\title{Calibrate probability by k-NN}
\usage{
knnCalib(
  y,
  trnScores,
  tstScores,
  k,
  loclin = FALSE,
  scaleX = NULL,
  smoothingFtn = NULL
)
}
\arguments{
\item{y}{R factor of labels in training set}

\item{trnScores}{vector/matrix of scores in training set}

\item{tstScores}{scores of new case(s)}

\item{k}{number of nearest neighbors}

\item{loclin}{indicate whether local linear regression should be applied}

\item{smoothingFtn}{a smoothing function whether it should be applied.}
}
\value{
vector of estimated probabilities for the new cases
}
\description{
for a given new case, the k nearest neighbors in the training set are
obtained; the class labels for the neighbors are then obtained, and
the resulting proportions for the different labels are then used as
the estimated probabilities for the binary class.
}
