% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{logOddsToProbs}
\alias{logOddsToProbs}
\title{Fit a logistic model}
\usage{
logOddsToProbs(x)
}
\arguments{
\item{x}{the input data}
}
\value{
the y output by a logistic model
}
\description{
Convert solved x to probability via logistic model
}
