% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{makeANDconjunction}
\alias{makeANDconjunction}
\title{replace old delimiter by 'AND' in the colnames of the decision values}
\usage{
makeANDconjunction(dvals, oldDelim)
}
\arguments{
\item{dvals}{column names with delimiters}

\item{oldDelim}{a delimiter}
}
\value{
column names of pairs of features to have 'AND' as delimiter
}
\description{
need column names of pairs of features to have 'AND' as delimiter,
e.g. 'CatcherANDFirst_Baseman' in the mlb data, replacing the original
delimiter, such as '/ for e1071
replace old delimiter by 'AND' in the colnames of the decision values
}
