% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{ovaCalib}
\alias{ovaCalib}
\title{Calibration for one-vs-all case}
\usage{
ovaCalib(
  trnY,
  trnScores,
  tstScores,
  calibMethod,
  deg = NULL,
  se = FALSE,
  K = NULL,
  scaleX = NULL,
  smoothingFtn = NULL,
  isProb = FALSE,
  fitAllX = FALSE
)
}
\arguments{
\item{trnY}{the training labels}

\item{trnScores}{a matrix of decision values or probabilities given by the classifier}

\item{tstScores}{a matrix of decision values or probabilities given by the classifier}

\item{calibMethod}{a string that indicate which calibration method to use.
options: "knnCalib", "locLinknnCalib", "plattCalib", "isoCalib", "BBQCalib", "ELiTECalib"}

\item{deg}{an integer to indicate the degree of polynomial for platt scaling}

\item{K}{an integer to indicate the number of nearest neighbor}

\item{scaleX}{an integer to indicate the number of nearest neighbor}

\item{smoothingFtn}{a smoothing function whether it should be applied}

\item{isProb}{indicate whether the scores are probabilities}

\item{fitAllX}{to determine whether you want to use all scores for fitting algorithms that are not knn-based.}
}
\value{
matrix of estimated probabilities for the new cases
}
\description{
a wrapper to implement calibration methods via one-vs-all approach
}
