% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{plattCalib}
\alias{plattCalib}
\title{Calibrate probability based on Platt scaling}
\usage{
plattCalib(trnY, trnScores, tstScores, deg, se = FALSE)
}
\arguments{
\item{trnY}{the training labels}

\item{trnScores}{a matrix of decision values or probabilities given by the classifier for the training sample}

\item{tstScores}{a matrix of decision values or probabilities given by the classifier for the test sample}

\item{deg}{an integer to indicate the degree of polynomial for platt scaling}

\item{se}{to determine whether standard errors should be computed via delta method}
}
\value{
a list that contains the following 1. prob: a vector of probabilities  2. se: a vector of standard errors if se is set to TRUE
}
\description{
apply ploynomial platt scaling to the posterior probabilities or decision values from SVM or other algorithms
}
