% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{preCalibWrap}
\alias{preCalibWrap}
\title{Prepare the data the for calibwrap}
\usage{
preCalibWrap(dta, yName, qeFtn = "qeSVM", qeArgs = NULL, holdout = 500)
}
\arguments{
\item{dta}{a dataframe for the input data}

\item{yName}{give the column name of the label in the data}

\item{qeFtn}{select the desired qe function}

\item{holdout}{the number of sample being held out}
}
\value{
a dataframe that lists all metrics for the given test labels and the calibrated probability matrix
}
\description{
preCalibWrap can be used on the orginal dataset, to set the holdout
}
