% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{predictPlatt}
\alias{predictPlatt}
\title{Predict the new case based on a trained logistic model}
\usage{
predictPlatt(prePlattCalibOut, tstScores, se = FALSE)
}
\arguments{
\item{prePlattCalibOut}{the logisitc model produced by fitPlatt()}

\item{tstScores}{a vector of decision values or probabilities given by the classifier for the testing sample}

\item{se}{to determine whether we want to compute the standard error through the delta method}
}
\value{
a list that contains the following 1. prob: a vector of probabilities; 2. se: a vector of standard errors if se is set to TRUE
}
\description{
Taking the logistic model and predict on the test scores given the classifier
}
