% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{reliabDiagram}
\alias{reliabDiagram}
\title{Reliability diagram and associated compilation of statistics}
\usage{
reliabDiagram(
  y,
  probs,
  nBins,
  plotGraph = TRUE,
  zoom = NULL,
  classNum = NULL,
  multiclass = FALSE,
  multiclassStyle = 1
)
}
\arguments{
\item{y}{vector of Y values from training set, 0s and 1s}

\item{probs}{vector of estimated cond. class probabilities}

\item{nBins}{number of bins}

\item{plotGraph}{TRUE means plotting is desired}

\item{zoom}{a tuple of values to set the limit of the axis e.g. (0, 100)}

\item{classNum}{an integer to specify the class number in the title of the plot}

\item{multiclass}{determine whether it is a multiclass case.}

\item{multiclassStyle}{1 or 2 to denote which multicalss reliability diagram style you prefer}
}
\value{
one reliability diagram
}
\description{
Draw the reliability diagram
}
