% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.r
\name{ROC}
\alias{ROC}
\title{Plot ROC}
\usage{
ROC(y, scores)
}
\arguments{
\item{y}{labels in training set; a 0s and 1s vector}

\item{scores}{values that your ML algorithm predicts from}
}
\value{
a ROC plot
}
\description{
Plot ROC
}
